% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrnstat.R
\name{rtrnstat}
\alias{rtrnstat}
\title{Compute re-transformed regression fit statistics}
\usage{
rtrnstat(regObj, tranPwr = 0)
}
\arguments{
\item{regObj}{A lm object}

\item{tranPwr}{Power transformation applied to the original variable of interest. Defaults to 0 (log).}
}
\description{
For a regression where the dependent variable is a power transformation of the original
variable of interest, standard error-like and R-squared-like statistics are computed
in the scale of the original variable of interest.
}
\examples{
fm1 <- lm(log(Volume) ~ log(Girth), data = trees); rtrnstat(fm1,0)
}

